removed code that was not being used, changed some naming conventions, refactored some of the material-ui calls
