+++ gracefabrics/src/App.css
@@ -1,45 +1,5 @@
 .App {
   text-align: center;
-
 }
 
 
-/* https://ak.picdn.net/shutterstock/videos/15516298/thumb/1.jpg
-https://c4.wallpaperflare.com/wallpaper/597/716/1020/purple-abstract-background-purple-and-white-wallpaper-wallpaper-preview.jpg
-https://www.zingerbug.com/Backgrounds/background_images/purple_marble_tie_dye_seamless.jpg
-https://i.pinimg.com/originals/e2/ce/da/e2cedad018d01878434ece8e70bd50f9.jpg
-*/
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
