+++ gracefabrics/src/components/UpdateForm/index.js
@@ -1,6 +1,5 @@
 import React from "react";
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
+import {Button, TextField} from '@material-ui/core';
 import './updateForm.css'
 import API from '../../utils/API';
